pyproject.toml
# LICENSE HEADER

[tool.poetry]
name = "app"
version = "1.0"
description = "App"
authors = ["Kuldeep K"]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.104.1"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0"
pytest = "^7.3.2"
httpx = "^0.25.2"
sqlalchemy = "2.0.41"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.0.0"
aiokafka = "^0.10.0"
pytest-asyncio = "^0.23.0"

[tool.poetry.group.dev.dependencies]
black = "24.3.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
